requirements.txt
Django==2.1.5
django-appconf==1.0.3
django-imagekit==4.0.2
djangorestframework==3.9.2
pilkit==2.0
Pillow==6.0.0
psycopg2-binary==2.8.2
pytz==2018.9
six==1.12.0
